backend/requirements.txt
flask
flask_sqlalchemy
flask_migrate
flask_cors
pyinstaller
